Adjust timeouts
