+++ packages/genesys-web-messaging-tester-cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ovotech/genesys-web-messaging-tester-cli",
-  "version": "1.0.13",
+  "version": "1.0.13-beta.2",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "license": "Apache-2.0",
@@ -27,7 +27,7 @@
     "web-messaging-tester": "lib/index.js"
   },
   "dependencies": {
-    "@ovotech/genesys-web-messaging-tester": "^2.0.5",
+    "@ovotech/genesys-web-messaging-tester": "2.0.5-beta.2",
     "chalk": "^4.1.2",
     "ci-info": "^3.5.0",
     "commander": "^11.0.0",
